+++ lib/app_module.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:hermes_app/home/home_module.dart';
 import 'package:hermes_app/shared/database/idatabase.dart';
-import 'package:hermes_app/shared/models/unknown_error.dart';
 import 'package:hermes_app/shared/repositories/category_repository.dart';
+import 'package:hermes_app/shared/widgets/default_error_widget/register_error_cubit.dart';
 
 class AppModule extends Module {
   @override
