+++ lib/shared/models/unknown_error.dart
@@ -1,8 +1,8 @@
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:hermes_app/shared/models/failure.dart';
 
-class RegisterErrorCubit extends Cubit {
-  RegisterErrorCubit(super.initialState);
-
-  void registerError(Failure error) {}
+class UnknownError extends Failure {
+  UnknownError({
+    required super.error,
+    super.stackTrace,
+  });
 }
