+++ lib/shared/widgets/default_error_widget/default_error_widget.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:hermes_app/shared/extensions/build_context_extensions.dart';
 import 'package:hermes_app/shared/models/failure.dart';
-import 'package:hermes_app/shared/models/unknown_error.dart';
 import 'package:hermes_app/shared/theme/app_colors.dart';
 import 'package:hermes_app/shared/widgets/default_button/default_button.dart';
+import 'package:hermes_app/shared/widgets/default_error_widget/register_error_cubit.dart';
 
 class DefaultErrorWidget extends StatefulWidget {
   final String title;
